start es in background

+++ src/snovault/tests/serverfixtures.py
@@ -1,4 +1,5 @@
 import pytest
+import os
 
 
 def pytest_configure():
@@ -68,11 +69,12 @@ def elasticsearch_server(request, elasticsearch_host_port):
     host, port = elasticsearch_host_port
     tmpdir = request.config._tmpdirhandler.mktemp('elasticsearch', numbered=True)
     tmpdir = str(tmpdir)
-    process = server_process(str(tmpdir), host=host, port=9200)
+    if not os.environ.get('TRAVIS'):
+        process = server_process(str(tmpdir), host=host, port=9200)
     print('PORT CHANGED')
     yield 'http://%s:%d' % (host, 9200)
 
-    if process.poll() is None:
+    if process and process.poll() is None:
         process.terminate()
         process.wait()
